n708-feedback/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
flask-jwt-extended==4.3.1
Werkzeug==2.3.3
python-dotenv==1.1.0